chore: Sets VisualStudioExtension .NET Framework version back to 4.6.1
This reverts commit 8cc858bf11f916ee336a60350906a1e6efd29655:
"chore: Sets VisualStudioExtension .NET Framework version to 4.7.2"

We switched to .NET 4.7.2 because building Sharpen in VS2019 was not
possible on 4.6.1. This is why:

The primary reference "Microsoft.VisualStudio.LanguageServices,
Version=2.4.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35,
processorArchitecture=MSIL" could not be resolved because it has an
indirect dependency on the assembly "Microsoft.Build.Framework,
Version=15.1.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"
which was built against the ".NETFramework,Version=v4.7.2" framework.
This is a higher version than the currently targeted framework
".NETFramework,Version=v4.6.1".

Anyhow, the issue needs a bit of investigation. For the upcoming
v0.9.0 release we want to avoid potential issues with the switch
to .NET 4.7.2. Yes, just being paranoid :-)